float split

+++ sdss_hist.py
@@ -42,7 +42,7 @@ if __name__ == "__main__":
                         help="List of columns")
     parser.add_argument("-galactic",  dest="gal_coords",action='store_true', default=False, required=False)
     parser.add_argument("-height",  dest="plot_height",action='store_true', default=False, required=False)
-    parser.add_argument("-split_Fe",  dest="split",type=int, required=False)
+    parser.add_argument("-split_Fe",  dest="split",type=float, required=False)
                         
     args = parser.parse_args()
 
@@ -81,4 +81,4 @@ if __name__ == "__main__":
    
 
     #Pass the columns to be plotted
-    Sdssplots.easy_hist(rtn_col_list, args.col_list, args.split = -999)
+    Sdssplots.easy_hist(rtn_col_list, args.col_list, args.split = -999.)
